Add a functionality to add new TODO items
Created a functionality that add a TODO item to the li when a text is entered to the input. Used append() to append a html string to the element we selected. And adding ul $("ul").on("click") with a second argument of "li" that specifies lis that may or may not have been on the page when it loaded inside the ul that was already exist when the page loaded